Move conversion function to conversion helper file
