+++ tango_ros_common/tango_ros_native/CMakeLists.txt
@@ -46,6 +46,7 @@ include_directories(
 )
 
 add_library(tango_ros_native SHARED
+  src/tango_ros_conversion.cpp
   src/tango_ros_node.cpp
   )
